+++ packages/client/src/helpers/__tests__/DynascaleManager.test.ts
@@ -9,7 +9,7 @@ import { DynascaleManager } from '../DynascaleManager';
 import { Call } from '../../Call';
 import { StreamClient } from '../../coordinator/connection/client';
 import { StreamVideoWriteableStateStore } from '../../store';
-import { DebounceType, VisibilityState } from '../../types';
+import { VisibilityState } from '../../types';
 import { noopComparator } from '../../sorting';
 import { TrackType } from '../../gen/video/sfu/models/models';
 
@@ -25,7 +25,7 @@ describe('DynascaleManager', () => {
       clientStore: new StreamVideoWriteableStateStore(),
     });
     call.setSortParticipantsBy(noopComparator());
-    dynascaleManager = new DynascaleManager(call);
+    dynascaleManager = new DynascaleManager(call.state, call.speaker);
   });
 
   afterEach(() => {
@@ -194,7 +194,10 @@ describe('DynascaleManager', () => {
     });
 
     it('video: should update subscription when track becomes available', () => {
-      const updateSubscription = vi.spyOn(call, 'updateSubscriptionsPartial');
+      const updateSubscription = vi.spyOn(
+        call.state,
+        'updateParticipantTracks',
+      );
 
       // @ts-ignore
       call.state.updateOrAddParticipant('session-id', {
@@ -213,52 +216,45 @@ describe('DynascaleManager', () => {
       expect(videoElement.muted).toBe(true);
       expect(videoElement.playsInline).toBe(true);
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
 
       call.state.updateParticipant('session-id', {
         publishedTracks: [TrackType.VIDEO],
       });
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        {
-          'session-id': {
-            dimension: {
-              width: videoElement.clientWidth,
-              height: videoElement.clientHeight,
-            },
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': {
+          dimension: {
+            width: videoElement.clientWidth,
+            height: videoElement.clientHeight,
           },
         },
-        DebounceType.FAST,
-      );
+      });
 
       call.state.updateParticipant('session-id', {
         publishedTracks: [TrackType.VIDEO],
       });
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
 
       cleanup?.();
 
-      expect(updateSubscription).toHaveBeenLastCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenLastCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
     });
 
     it('video: should play video when track becomes available', () => {
       vi.useFakeTimers();
 
-      const updateSubscription = vi.spyOn(call, 'updateSubscriptionsPartial');
+      const updateSubscription = vi.spyOn(
+        call.state,
+        'updateParticipantTracks',
+      );
       const play = vi.spyOn(videoElement, 'play').mockResolvedValue();
 
       // @ts-ignore
@@ -282,28 +278,22 @@ describe('DynascaleManager', () => {
 
       vi.runAllTimers();
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        {
-          'session-id': {
-            dimension: {
-              width: videoElement.clientWidth,
-              height: videoElement.clientHeight,
-            },
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': {
+          dimension: {
+            width: videoElement.clientWidth,
+            height: videoElement.clientHeight,
           },
         },
-        DebounceType.FAST,
-      );
+      });
       expect(play).toHaveBeenCalled();
       expect(videoElement.srcObject).toBe(mediaStream);
 
       cleanup?.();
 
-      expect(updateSubscription).toHaveBeenLastCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenLastCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
     });
 
     it('video: should update subscription when element becomes visible', () => {
@@ -318,7 +308,10 @@ describe('DynascaleManager', () => {
         },
       });
 
-      const updateSubscription = vi.spyOn(call, 'updateSubscriptionsPartial');
+      const updateSubscription = vi.spyOn(
+        call.state,
+        'updateParticipantTracks',
+      );
 
       const cleanup = dynascaleManager.bindVideoElement(
         videoElement,
@@ -326,11 +319,9 @@ describe('DynascaleManager', () => {
         'videoTrack',
       );
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
 
       call.state.updateParticipant('session-id', {
         viewportVisibilityState: {
@@ -339,18 +330,14 @@ describe('DynascaleManager', () => {
         },
       });
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        {
-          'session-id': {
-            dimension: {
-              width: videoElement.clientWidth,
-              height: videoElement.clientHeight,
-            },
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': {
+          dimension: {
+            width: videoElement.clientWidth,
+            height: videoElement.clientHeight,
           },
         },
-        DebounceType.MEDIUM,
-      );
+      });
 
       call.state.updateParticipant('session-id', {
         viewportVisibilityState: {
@@ -359,11 +346,9 @@ describe('DynascaleManager', () => {
         },
       });
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.MEDIUM,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
 
       call.state.updateParticipant('session-id', {
         viewportVisibilityState: {
@@ -372,26 +357,20 @@ describe('DynascaleManager', () => {
         },
       });
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        {
-          'session-id': {
-            dimension: {
-              width: videoElement.clientWidth,
-              height: videoElement.clientHeight,
-            },
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': {
+          dimension: {
+            width: videoElement.clientWidth,
+            height: videoElement.clientHeight,
           },
         },
-        DebounceType.MEDIUM,
-      );
+      });
 
       cleanup?.();
 
-      expect(updateSubscription).toHaveBeenLastCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenLastCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
     });
 
     it('video: should update subscription when element resizes', () => {
@@ -406,7 +385,7 @@ describe('DynascaleManager', () => {
         },
       });
 
-      let updateSubscription = vi.spyOn(call, 'updateSubscriptionsPartial');
+      let updateSubscription = vi.spyOn(call.state, 'updateParticipantTracks');
 
       let resizeObserverCallback: ResizeObserverCallback;
       window.ResizeObserver = class ResizeObserver {
@@ -428,18 +407,14 @@ describe('DynascaleManager', () => {
         'videoTrack',
       );
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        {
-          'session-id': {
-            dimension: {
-              width: videoElement.clientWidth,
-              height: videoElement.clientHeight,
-            },
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': {
+          dimension: {
+            width: videoElement.clientWidth,
+            height: videoElement.clientHeight,
           },
         },
-        DebounceType.FAST,
-      );
+      });
 
       // @ts-ignore simulate resize
       videoElement.clientHeight = 101;
@@ -449,19 +424,15 @@ describe('DynascaleManager', () => {
       // @ts-ignore simulate resize
       resizeObserverCallback();
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: { width: 101, height: 101 } } },
-        DebounceType.SLOW,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: { width: 101, height: 101 } },
+      });
 
       cleanup?.();
 
-      expect(updateSubscription).toHaveBeenLastCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenLastCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
     });
 
     it('video: should unsubscribe when element dimensions are zero', () => {
@@ -476,7 +447,7 @@ describe('DynascaleManager', () => {
         },
       });
 
-      let updateSubscription = vi.spyOn(call, 'updateSubscriptionsPartial');
+      let updateSubscription = vi.spyOn(call.state, 'updateParticipantTracks');
 
       // @ts-ignore simulate resize
       videoElement.clientHeight = 0;
@@ -489,19 +460,15 @@ describe('DynascaleManager', () => {
         'videoTrack',
       );
 
-      expect(updateSubscription).toHaveBeenCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
 
       cleanup?.();
 
-      expect(updateSubscription).toHaveBeenLastCalledWith(
-        'videoTrack',
-        { 'session-id': { dimension: undefined } },
-        DebounceType.FAST,
-      );
+      expect(updateSubscription).toHaveBeenLastCalledWith('videoTrack', {
+        'session-id': { dimension: undefined },
+      });
     });
   });
 });