+++ packages/react-sdk/src/core/components/Video/Video.tsx
@@ -21,6 +21,17 @@ import { useCall } from '@stream-io/video-react-bindings';
 
 export type VideoProps = ComponentPropsWithoutRef<'video'> & {
   /**
+   * Pass false to disable rendering video and render fallback
+   * even if the participant has published video.
+   * @default true
+   */
+  enabled?: boolean;
+  /**
+   * Forces the video to be mirrored or unmirrored. By default, video track
+   * from the local participant is mirrored, and all other videos are not mirrored.
+   */
+  mirror?: boolean;
+  /**
    * The track type to display.
    */
   trackType: VideoTrackType | 'none';
@@ -54,6 +65,8 @@ export type VideoProps = ComponentPropsWithoutRef<'video'> & {
 };
 
 export const Video = ({
+  enabled = true,
+  mirror,
   trackType,
   participant,
   className,
@@ -139,8 +152,11 @@ export const Video = ({
     trackType === 'none' ||
     viewportVisibilityState?.[trackType] === VisibilityState.INVISIBLE;
 
-  const hasNoVideoOrInvisible = !isPublishingTrack || isInvisible;
-  const mirrorVideo = isLocalParticipant && trackType === 'videoTrack';
+  const hasNoVideoOrInvisible = !enabled || !isPublishingTrack || isInvisible;
+  const mirrorVideo =
+    mirror === undefined
+      ? isLocalParticipant && trackType === 'videoTrack'
+      : mirror;
   const isScreenShareTrack = trackType === 'screenShareTrack';
   return (
     <>