+++ sample-apps/react/react-dogfood/components/ActiveCall.tsx
@@ -27,6 +27,7 @@ import { ChatUI } from './ChatUI';
 import { CallStatsSidebar, ToggleStatsButton } from './CallStatsWrapper';
 import { ClosedCaptions, ClosedCaptionsSidebar } from './ClosedCaptions';
 import { ToggleSettingsTabModal } from './Settings/SettingsTabModal';
+import { IncomingVideoSettingsButton } from './IncomingVideoSettings';
 import { ToggleEffectsButton } from './ToggleEffectsButton';
 import { ToggleNoiseCancellationButton } from './ToggleNoiseCancellationButton';
 import { ToggleFeedbackButton } from './ToggleFeedbackButton';
@@ -261,6 +262,11 @@ export const ActiveCall = (props: ActiveCallProps) => {
             <div className="str-video__call-controls__desktop">
               <ScreenShareButton />
             </div>
+            {isPronto && (
+              <div className="str-video__call-controls__desktop">
+                <IncomingVideoSettingsButton />
+              </div>
+            )}
             <RecordCallConfirmationButton />
             <div className="str-video__call-controls__desktop">
               <CancelCallConfirmButton onLeave={onLeave} />